// frontend/types/designTypeEnum.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * My API
 * Documentação automática da API
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `full_factorial` - Full Factorial
* `fractional_factorial` - Fractional Factorial
* `plackett_burman` - Plackett-Burman
* `box_behnken` - Box-Behnken
* `central_composite` - Central Composite
 */
export type DesignTypeEnum = typeof DesignTypeEnum[keyof typeof DesignTypeEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DesignTypeEnum = {
  full_factorial: 'full_factorial',
  fractional_factorial: 'fractional_factorial',
  plackett_burman: 'plackett_burman',
  box_behnken: 'box_behnken',
  central_composite: 'central_composite',
} as const;