// frontend/types/statusEnum.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * My API
 * Documentação automática da API
 * OpenAPI spec version: 1.0.0
 */

/**
 * * `draft` - Draft
* `design_ready` - Design Ready
* `data_collection` - Data Collection
* `analysis_ready` - Analysis Ready
* `completed` - Completed
* `archived` - Archived
 */
export type StatusEnum = typeof StatusEnum[keyof typeof StatusEnum];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const StatusEnum = {
  draft: 'draft',
  design_ready: 'design_ready',
  data_collection: 'data_collection',
  analysis_ready: 'analysis_ready',
  completed: 'completed',
  archived: 'archived',
} as const;